requirements.txt
asgiref==3.8.1
channels==4.0.0
channels-redis==4.2.0
crispy-bootstrap5
Django==5.0.7
django-crispy-forms
dj-database-url==2.1.0  # <-- Добавили
gunicorn==22.0.0
psycopg2-binary
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.7.0















